refactor: GameConstants.ts を GameConfig.ts に統合
